fix(skill): open-spec 合规审计修复 v0.3.1
- SKILL.md: "重要约束"改为"严格禁止"，使用 Authority 原则措辞
- SKILL.md: 核心工作流写入步骤添加 HARD-GATE 标注
- SKILL.md: 意图判断新增关键区分(在线表格/多维表/文档)
- package.json: description 同步 frontmatter，补充同义词和排除场景
- package.json: 版本号升级至 0.3.1
- CHANGELOG.md: 记录 v0.3.1 变更
