+++ CHANGELOG.md
@@ -1,5 +1,53 @@
 # Changelog
 
+## [0.3.1] - 2026-03-05
+
+### 修复 (open-spec 合规审计)
+
+**SKILL.md 三项审计修复：**
+- ✅ "重要约束"改为"严格禁止"，使用 Authority 原则措辞（"禁止"/"必须"）
+- ✅ 核心工作流写入步骤添加 HARD-GATE 标注 + "提取什么"标注
+- ✅ 意图判断新增"关键区分"说明，明确区分在线表格/多维表/文档
+
+**package.json description 同步：**
+- ✅ description 与 SKILL.md frontmatter 保持一致，补充口语同义词和排除场景
+
+---
+
+## [0.3.0] - 2026-03-05
+
+### 重构
+
+**SKILL.md 按 dingtalk-neulink 技能包规范全面重构：**
+- ✅ 满足 7 区块规范（frontmatter / 严格禁止 / 方法总览 / 意图决策树 / 核心工作流 / 上下文传递 / 错误处理）
+- ✅ 主文件控制在 117 行（规范要求 ≤ 150 行）
+- ✅ 严格禁止 7 条，含产品专属禁令（类型混淆、URL 格式、覆盖写入确认等）
+- ✅ 意图决策树覆盖 6 种用户意图，使用中文自然语言 + 口语化表达
+- ✅ 上下文传递链完整无断裂
+- ✅ 参数格式区块含正确/错误 JSON 对比示例
+
+**package.json 规范化：**
+- ✅ description 符合 D-01~D-08 规则（动词短语开头、含触发场景、含排除场景、含口语同义词、无内部术语）
+- ✅ 版本号升级至 0.3.0
+- ✅ keywords 移除 "mcp"，新增 "钉钉文档"、"在线文档"
+
+**references/api-reference.md 增强：**
+- ✅ 每个方法描述包含四要素（一句话用途 + 使用场景 + 区分说明 + 调用示例）
+- ✅ 参数来源标注（如 parentDentryUuid 来自 get_my_docs_root_dentry_uuid）
+- ✅ 类型陷阱高亮（accessType 字符串、updateType 数字、docUrl 完整 URL）
+
+**references/error-codes.md 精简：**
+- ✅ 错误码速查表精简为一行一错误，含具体修复动作
+- ✅ 新增参数类型速查表（正确/错误/后果）
+- ✅ 调试流程精简为 6 步
+
+**README.md 同步更新：**
+- ✅ 使用示例改为 `--args` JSON 格式
+- ✅ 新增目录结构说明
+- ✅ 新增注意事项（参数类型陷阱）
+
+---
+
 ## [0.2.1] - 2026-03-04
 
 ### 修复
