+++ package.json
@@ -1,13 +1,14 @@
 {
   "name": "dingtalk-docs",
-  "version": "0.2.1",
-  "description": "钉钉文档操作技能。使用 mcporter CLI 连接钉钉 MCP server 执行文档搜索、创建、内容读写等操作。提供 6 个真实可用的 API 方法。",
+  "version": "0.3.1",
+  "description": "管理钉钉云文档中的文档、文件夹和内容。当用户想要创建文档、搜索文档、读取或写入文档内容、创建文件夹整理文档时使用。也适用于用户提到云文档、在线文档、钉钉文档、钉文档等关键词的场景。不要在用户需要操作多维表、管理日程、发消息或处理审批流时触发。",
   "keywords": [
     "dingtalk",
     "docs",
     "document",
-    "mcp",
     "云文档",
+    "钉钉文档",
+    "在线文档",
     "openclaw",
     "skill"
   ],
@@ -37,7 +38,7 @@
     "credentials": [
       {
         "name": "DINGTALK_MCP_DOCS_URL",
-        "description": "钉钉文档 MCP Server Streamable HTTP URL (含访问令牌)",
+        "description": "钉钉文档服务的 Streamable HTTP 地址（含访问令牌）",
         "docs": "https://mcp.dingtalk.com",
         "storageMethod": "mcporter config (recommended) or environment variable"
       }
