+++ scripts/import_docs.py
@@ -15,12 +15,10 @@
 """
 
 import sys
-import subprocess
-import os
-import re
-import json
 from pathlib import Path
-from typing import Optional, Tuple
+from typing import Optional
+
+from mcporter_utils import run_mcporter, parse_response, get_root_dentry_uuid, resolve_safe_path
 
 # ============== 安全常量 ==============
 MAX_FILE_SIZE = 10 * 1024 * 1024  # 10MB
@@ -28,26 +26,6 @@ MAX_CONTENT_LENGTH = 50000  # 最大内容长度
 ALLOWED_EXTENSIONS = ['.md', '.txt', '.markdown']
 
 # ============== 安全函数 ==============
-
-def resolve_safe_path(path: str) -> Path:
-    """解析路径并限制在工作目录内"""
-    allowed_root = os.environ.get('OPENCLAW_WORKSPACE', os.getcwd())
-    allowed_root = Path(allowed_root).resolve()
-
-    if Path(path).is_absolute():
-        target_path = Path(path).resolve()
-    else:
-        target_path = (Path.cwd() / path).resolve()
-
-    try:
-        target_path.relative_to(allowed_root)
-        return target_path
-    except ValueError:
-        raise ValueError(
-            f"路径超出允许范围：{path}\n"
-            f"允许根目录：{allowed_root}\n"
-            f"提示：设置 OPENCLAW_WORKSPACE 环境变量"
-        )
 
 def validate_file_extension(filename: str) -> bool:
     """验证文件扩展名"""
@@ -61,52 +39,6 @@ def validate_file_size(path: Path) -> bool:
         print(f"❌ 文件过大：{size / 1024 / 1024:.2f}MB（最大 {MAX_FILE_SIZE / 1024 / 1024}MB）")
         return False
     return True
-
-# ============== 工具函数 ==============
-
-def run_mcporter(tool: str, args: dict = None, timeout: int = 60) -> Tuple[bool, str]:
-    """执行 mcporter 命令（使用 --args JSON 传参）"""
-    command = ['mcporter', 'call', tool, '--output', 'json']
-    if args:
-        command.extend(['--args', json.dumps(args, ensure_ascii=False)])
-    try:
-        result = subprocess.run(
-            command,
-            capture_output=True,
-            text=True,
-            timeout=timeout
-        )
-        if result.returncode == 0:
-            return True, result.stdout
-        else:
-            return False, result.stderr
-    except subprocess.TimeoutExpired:
-        return False, f"命令执行超时（{timeout}秒）"
-    except Exception as e:
-        return False, str(e)
-
-def parse_response(output: str) -> Optional[dict]:
-    """解析 mcporter 响应，自动处理嵌套 result 结构"""
-    try:
-        data = json.loads(output)
-        if isinstance(data, dict) and 'result' in data:
-            return data['result']
-        return data
-    except json.JSONDecodeError:
-        return None
-
-def get_root_dentry_uuid() -> Optional[str]:
-    """获取根目录 ID"""
-    success, output = run_mcporter('dingtalk-docs.get_my_docs_root_dentry_uuid')
-
-    if not success:
-        print(f"❌ 获取根目录 ID 失败：{output}")
-        return None
-
-    result = parse_response(output)
-    if result is None:
-        return None
-    return result.get('rootDentryUuid')
 
 def create_doc(title: str, parent_uuid: str) -> Optional[str]:
     """创建文档"""
